Rename gitignore to .gitignore to avoid tracking unwanted files
