+++ src/holiday_budget_planner/models/__init__.py
@@ -1,0 +1,5 @@
+"""Data models used across the holiday budget planner package."""
+
+from .expense import Expense
+
+__all__ = ["Expense"]
