+++ src/holiday_budget_planner/services/__init__.py
@@ -1,0 +1,4 @@
+from .expense_book import ExpenseBook, PEOPLE
+from . import charts
+
+__all__ = ["ExpenseBook", "PEOPLE", "charts"]
