+++ src/holiday_budget_planner/utils/__init__.py
@@ -1,0 +1,3 @@
+from .rounding import d_round
+
+__all__ = ["d_round"]
